Move event listing funcionality to GoogleEventLister class
